Data
I am getting the positions out from the script.
To Do:
Figure out how long it should run for
Write Script to turn positions into graphs
Figure out how to get it to give optimal speed.
Figure out how to get it to continue to go on a circular track